[STF][BUILD] Move Constants to STF-Self-Tests
I had to move the Constants.java class away from the stf-test because
they will be used by the stf-self-tests either. This causes all the new
imports.
