+++ de.fu_berlin.inf.dpp.stf.test/test/de/fu_berlin/inf/dpp/stf/test/consistency/ModifyDocumentBeforeProjectNegotiationTest.java
@@ -10,7 +10,7 @@ import static org.junit.Assert.assertEquals;
 
 import de.fu_berlin.inf.dpp.stf.client.StfTestCase;
 import de.fu_berlin.inf.dpp.stf.client.util.Util;
-import de.fu_berlin.inf.dpp.stf.test.Constants;
+import de.fu_berlin.inf.dpp.stf.test.stf.Constants;
 import org.junit.BeforeClass;
 import org.junit.Test;
