+++ de.fu_berlin.inf.dpp/test/framework/stf/test/de/fu_berlin/inf/dpp/stf/test/stf/contextmenu/ContextMenuDeleteTest.java
@@ -3,7 +3,7 @@ package de.fu_berlin.inf.dpp.stf.test.stf.contextmenu;
 import static de.fu_berlin.inf.dpp.stf.client.tester.SarosTester.ALICE;
 
 import de.fu_berlin.inf.dpp.stf.client.StfTestCase;
-import de.fu_berlin.inf.dpp.stf.test.Constants;
+import de.fu_berlin.inf.dpp.stf.test.stf.Constants;
 import java.rmi.RemoteException;
 import org.junit.After;
 import org.junit.BeforeClass;
