+++ de.fu_berlin.inf.dpp/test/framework/stf/test/de/fu_berlin/inf/dpp/stf/test/stf/contextmenu/ContextMenuShareWithTest.java
@@ -5,7 +5,7 @@ import static de.fu_berlin.inf.dpp.stf.client.tester.SarosTester.BOB;
 
 import de.fu_berlin.inf.dpp.stf.client.StfTestCase;
 import de.fu_berlin.inf.dpp.stf.shared.Constants.TypeOfCreateProject;
-import de.fu_berlin.inf.dpp.stf.test.Constants;
+import de.fu_berlin.inf.dpp.stf.test.stf.Constants;
 import org.junit.After;
 import org.junit.BeforeClass;
 import org.junit.Test;
