+++ de.fu_berlin.inf.dpp/test/framework/stf/test/de/fu_berlin/inf/dpp/stf/test/stf/editor/EditorByAliceBobTest.java
@@ -8,7 +8,7 @@ import static org.junit.Assert.assertTrue;
 
 import de.fu_berlin.inf.dpp.stf.client.StfTestCase;
 import de.fu_berlin.inf.dpp.stf.client.util.Util;
-import de.fu_berlin.inf.dpp.stf.test.Constants;
+import de.fu_berlin.inf.dpp.stf.test.stf.Constants;
 import java.rmi.RemoteException;
 import org.junit.Before;
 import org.junit.BeforeClass;
