+++ de.fu_berlin.inf.dpp/test/framework/stf/test/de/fu_berlin/inf/dpp/stf/test/stf/view/explorer/PackageExplorerViewTest.java
@@ -6,7 +6,7 @@ import static org.junit.Assert.assertFalse;
 import static org.junit.Assert.assertTrue;
 
 import de.fu_berlin.inf.dpp.stf.client.StfTestCase;
-import de.fu_berlin.inf.dpp.stf.test.Constants;
+import de.fu_berlin.inf.dpp.stf.test.stf.Constants;
 import java.rmi.RemoteException;
 import java.util.regex.Pattern;
 import org.junit.After;
